+++ packages/headless/src/app/case-assist-thunk-extra-arguments.ts
@@ -1,0 +1,5 @@
+import {CaseAssistAPIClient} from '../api/service/case-assist/case-assist-api-client';
+import {ClientThunkExtraArguments} from './thunk-extra-arguments';
+
+export interface CaseAssistThunkExtraArguments
+  extends ClientThunkExtraArguments<CaseAssistAPIClient> {}